refactor: move editor components to under `/tasks`
