+++ src/components/ui/drawers/tasks/EditTaskDrawerContent.tsx
@@ -6,7 +6,7 @@ import {
 import { Input } from '@/components/ui/input';
 import { FormMessage } from '@/components/ui/formMessage';
 import { cn } from '@/lib/utils';
-import { TaskDescriptionEditor } from '@/components/ui/drawers/task-description/TaskDescriptionEditor';
+import { TaskDescriptionEditor } from '@/components/ui/drawers/tasks/task-description/TaskDescriptionEditor';
 import { Button } from '@/components/ui/button';
 import { CommonDrawer } from '@/components/ui/commons/CommonDrawer';
 import { useFormContext } from 'react-hook-form';
